requirements.txt
python-telegram-bot==20.3
flask==2.2.3
requests==2.28.1
beautifulsoup4==4.11.1
schedule==1.1.0